# docs/api_spec.yaml
components:
  schemas:
...
    SearchRequest:
      properties:
        context:
          properties:
            org_id:
              type: string
            user_id:
              type: string
          type: object
        limit:
          default: 10
          maximum: 100
          minimum: 1
          type: integer
        query:
          example: user preferences for dark theme
          type: string
      required:
      - query
      - context
      type: object
    SearchResponse:
      properties:
        query_time:
          example: 0.045
          type: number
        results:
          items:
            properties:
              memory_id:
                type: string
              score:
                maximum: 1
                minimum: 0
                type: number
            type: object
          type: array
        total:
          example: 25
          type: integer
      required:
      - results
      - total
      - query_time
      type: object
...
  securitySchemes:
    BearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http
info:
  contact:
    email: support@liftos.com
    name: Lift OS Core Team
  description: Unified API documentation for the Lift OS Core platform - a modular,
    scalable operating system for orchestrating all Lift products
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Lift OS Core API
  version: 1.0.0
openapi: 3.0.3
paths:
...
  /memory/search:
    post:
      description: Perform semantic search using the Knowledge Storage Engine
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
          description: Search completed successfully
        '400':
          description: Invalid search query
        '401':
          description: Unauthorized
      security:
      - BearerAuth: []
      summary: Semantic Search
      tags:
      - KSE Operations
...
servers:
- description: Development Gateway Server
  url: http://localhost:8000
- description: Production Gateway Server
  url: https://api.liftos.com